+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 * Add AUTHORITY type into folioRecordTypes (UIDATIMP-1021)
 * Adjust the FOLIO record type in the field mapping profile (UIDATIMP-1027)
 * Adjust the FOLIO record type in the action profile (UIDATIMP-1028)
+* Cover `<ViewContainer>` component with tests (UIDATIMP-731)
+* Cover `<ActionProfiles>` settings with tests (UIDATIMP-971)
+* Use correct `css-loader` syntax (UIDATIMP-1037)
 * Cover `<Jobs>` component with tests (UIDATIMP-959)
 
 ## [5.0.1](https://github.com/folio-org/ui-data-import/tree/v5.0.1) (2021-10-19)